Split imports to reduce chunk size
