+++ waitlst/src/components/LandingNavigationFeaturesContent.vue
@@ -1,6 +1,4 @@
 <script setup lang="ts">
-import { defineEmits } from 'vue'
-
 const emit = defineEmits(['closeMenu'])
 </script>
 <template>
