+++ waitlst/src/views/LandingView.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
-import LandingMainContent from '@/components/LandingMainContent.vue'
+import LandingContent from '@/components/LandingContent.vue'
 import LandingNavBar from '@/components/LandingNavBar.vue'
 </script>
 
 <template>
   <div class="grid gap-16 sm:gap-24 pb-16">
     <LandingNavBar />
-    <LandingMainContent />
+    <LandingContent />
   </div>
 </template>
